Refactor tests to be run on local dev: * Remove python client tests * Reuse sending csv capability from postman * Move user registration, invite and email template methods in util for re-use elsewhere
